TransactionService, removing trash files

+++ lib/model/account.ex
@@ -10,7 +10,7 @@ defmodule Account do
     field(:number, :integer, null: false)
     field(:agency, :integer, null: false)
     field(:currency, :string, default: "BRL")
-    field(:balance, :float, default: 0.0)
+    field(:balance, :integer, default: 0)
     timestamps()
   end
 end
